Move the 'GetStarted' pageobject to the proper folder
